+++ core/frontend/src/tile/internal.ts
@@ -43,10 +43,22 @@ export * from "./GltfReader";
 export * from "./I3dmReader";
 export * from "./B3dmReader";
 export * from "./ImdlReader";
+export * from "./map/ArcGisTokenGenerator";
 export * from "./map/ArcGisTokenManager";
 export * from "./map/MapLayerFormatRegistry";
+export * from "./map/ArcGisUtilities";
+export * from "./map/WmsUtilities";
+export * from "./map/WmsCapabilities";
+export * from "./map/WmtsCapabilities";
+export * from "./map/MapLayerImageryProvider";
+export * from "./map/ImageryProviders/ArcGISMapLayerImageryProvider";
+export * from "./map/ImageryProviders/AzureMapsLayerImageryProvider";
+export * from "./map/ImageryProviders/BingImageryProvider";
+export * from "./map/ImageryProviders/MapBoxLayerImageryProvider";
+export * from "./map/ImageryProviders/TileUrlImageryProvider";
+export * from "./map/ImageryProviders/WmsMapLayerImageryProvider";
+export * from "./map/ImageryProviders/WmtsMapLayerImageryProvider";
 export * from "./map/MapLayerImageryFormats";
-export * from "./map/BingImageryProvider";
 export * from "./map/MapLayerTileTreeReference";
 export * from "./map/MapTileTree";
 export * from "./map/TerrainMeshProvider";
@@ -67,10 +79,5 @@ export * from "./PrimaryTileTree";
 export * from "./ClassifierTileTree";
 export * from "./OrbitGtTileTree";
 export * from "./map/ImageryTileTree";
-export * from "./map/ArcGisUtilities";
-export * from "./map/ArcGisTokenGenerator";
-export * from "./map/WmsUtilities";
-export * from "./map/WmsCapabilities";
-export * from "./map/WmtsCapabilities";
 export * from "./map/MapLayerSources";
 export * from "./map/MapLayerSettingsService";
